Limit SBE Update Auto-Update of Minimum Secure Version to Production Drivers
This commit adds a check that the current driver is a production driver
before auto-updating the minimum secure version (aka MSV) of 0 to 1 even
if the Secure Version Lock-In policy is disabled.  The check will continue to
only work if the system is in securemode and the current MSV is 0.

Change-Id: Icf53d9d65ca1c4a0e3f98e0189f3d15db33d81aa
CQ:SW533842
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122167
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>