PRD: don't explicitly callout endpoints for ABUS callouts
For bus callouts, PRD will call the HWSV bus callout interface at low
priority, which will callout all parts on the bus including endpoints.
Then as a RAS policy, PRD will callout the endpoints at MED_A priority.
The endpoints will be called out twice, but only the highest priority
callout will be added to the error log. Now, in P10 with the new TDR
function, we have the ability to determine which part along the bus is
most likely at fault and TDR will adjust the priorities accordingly. The
problem is that this conflicts with the PRD endpoint callouts.
Therefore, PRD will not callout the endpoints for ABUS callouts and rely
on TDR to adjust the priorities accordingly. I have been assured by the
CBLV team that they will adjust the priorities to be consistent with
past RAS policies when TDR is unavailable (i.e. when there is a system
checkstop).

Change-Id: Ie5f5aa966f41979513a37b3306a796cfa6a53271
CQ: SW533595
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122228
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122295
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>