SBE dump request interface
This commit contains the beginning of the interface that Hostboot will
use to request that the BMC perform an SBE dump.

When Hostboot wants to ask the BMC to dump an SBE, it will make a PLDM
Set Numeric Effecter Value request, and the value will be the ID of
the error log that is responsible for starting the dump. Then Hostboot
will wait for the SBE dump to complete, which part will be implemented
in a later commit.

Change-Id: Ieac20bf63720e8fbb9ee72fd5a6e8e50c56697cc
RTC: 247294
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122255
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>