Fix host-initiated PLDM soft-poweroff flow
There are two flows for soft poweroffs depending on whether they are
initiated by the BMC or the host:

A. BMC-initiated
   1. BMC sends a PLDM request to the Host to shut down
   2. Host shuts down and notifies the BMC via PLDM
   3. The BMC cuts chassis power
B. Host-initiated
   1. The host shuts down
   2. The host sends a PLDM request to the BMC to cut chassis power

Previously, Hostboot conflated these two flows, and always sent a
shutdown notification rather than a chassis-off command to the BMC
when it was finished shutting down.

This commit splits the flows, so that the host will instruct the BMC
to cut power to the chassis when the host initiates the soft poweroff.

Change-Id: Ia734a52f53d7fda403af5a63b813826b6d450654
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122391
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>