PRD: limit fabric sorting FFDC
During checkstop analysis, PRD will query the global registers for each
processor to determine if the chip contains an ctive checkstop attention
or if the checkstop originated from another processor. There is some
code to check if the node is SMP coherent so that PRD only query
processors on a single node if the node has not been connected to the
other nodes. Unfortunately, this code is not used when gathering the
global register values for FFDC. So PRD will end up SCOMing the other
nodes, which could be in an invalid/unknown state. This commit will fix
that issue.

Change-Id: I9ed34ff1a15e0fc34deb67f834faa97967331291
CQ: SW533781
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122331
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122348
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>