Assume SBE ipl-time lid has Version Checking per EC
When we use PLDM for PNOR access Hostboot is not told whether or
not the lids we are using to back the PNOR data have sha512 hashes
for version checking or not. It is safe to assume that the SBE
lid will always have this so we will do so for configurations
where we use PLDM for pnor acess.

Change-Id: If346a08f0930169ca88b381a16c7684648120d34
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122402
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>