Skip polling SBE status register if we know the SBE is halted
There is a bug in the sbe retry handler where we will spend 1 minute
(10 minutes in simics) polling the sbe status register prior attempting
the retry, even if we know the SBE has been halted. This commit will
skip this initial polling loop if we pass the parameter to the retry
handler saying we know the SBE is halted.

Change-Id: Ie24a74537c5239ff8dfbd0b0c4fa7f945fceb0b1
CQ: SW534228
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122406
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>