Notify BMC with RX_COMPLETE prior to handling MCTP packet
When we have read a MCTP packet off of the RX buffer we should
immediately notify the BMC that we have done this. Previously
we would pass the packet into the mctp core logic prior to writing
the KCS interface. If the packet contained had EOM bit set, it would
send the complete message out to PLDM. In situations where we send
many PLDM requests in a row, we were seeing a race condition of who
set the KCS data register. If the TX_READY for the next PLDM message
was written prior to writing the RX_COMPLETE from the previous message
then we would enter a deadlock state. This change ensures that prior
to handling the contents of the MCTP packet we get, we notify the BMC
that we have read it.

Change-Id: I8421441de64a4269f33f8977ce085606e170143e
CQ: SW518989
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122436
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>