Complete implementation of SBE dump request
BACKGROUND:

On eBMC systems, when Hostboot attempts to boot the SBE in a secondary
processor and detects an error, it will ask the BMC to collect a dump
of the relevant SBE state for debugging purposes. It does this by
making a PLDM request to set a numeric effecter on the BMC, and then
waits until the BMC sets a state effecter on the host indicating that
the dump completed or failed.

THIS COMMIT:

This commit finalizes the implementation of the PLDM SBE dump
request. Also does some small refactoring to clean up relevant parts
of the PLDM codebase.

RTC: 247294
Change-Id: I68e9ebfe849da65f1327f33227b55c07af37c054
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122646
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>