Fix errlSrc section length to include FRU callout size
errlSrc had its section length to be hardcoded as 72 bytes plus 8 for
the header. This is no longer an accurate length because FRU callouts
can be in the primary SRC. So, when they are then the length needs to be
determined dynamically so that errlParser doesn't trip up when
processing all the existing sections in an error log.

Change-Id: Iaded4ecf128b8c5cf090627fdaa714a3ce9de7cb
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/122555
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>