Drain ODR prior to activating MCTP channel between Host/BMC
In an effort to clean up any stale effort left from previous
system states, Hostboot will drain the ODR prior to sending the
"INIT" KCS command to this BMC. Hostboot will drain the ODR by
reading the KCS status, checking if the OBF bit is set, if it is,
it reads the ODR and discards the message. Reading the ODR will
clear the OBF bit in the KCS status. Hostboot will then wait 50
msec and poll again. Hostboot will drain up to 255 messages from
the ODR before logging a predictive error and continuing on with
the init sequence. We hope this will fix some of the CRC errors
we have seen upon starting Hostboot's poll on the KCS status
register.

Change-Id: I0601c8f80cee442003272ba4454eaba91c03c564
CQ: SW534300
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123008
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123919
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>