Support for runtime PLDM SBE HRESET requests
This commit adds runtime support for PLDM SBE HRESET requests.

A PLDM terminus can send a SetStateEffecterStates request to the
Hostboot runtime to begin an HRESET operation. The runtime will
respond immediately and begin performing the HRESET operation. Once
done, the runtime will notify the BMC that the operation is complete
via a PLDM Sensor Changed event.

Hostboot will create a PLDM sensor that indicates the current status
of the HRESET operation in progress (one sensor per SBE). The sensor
should not be queried directly; the BMC should wait on Sensor State
Changed events (and cache the value from the event) to determine the
status of the sensor. (This is because currently HBRT cannot respond
to PLDM requests while an HRESET is in progress, and any request for
the sensor state will time out.)

Change-Id: Ic2d4255eb56bbd1da5bddf1ae0078f6730739247
RTC: 247294
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123080
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>