Add mutex protection around calls that write the IDR
Prior to this commit in the mctprp code we had mutex protection around
calls that would pontentially remove entries from the tx_queue because
we could have multiple tasks attempting to do this simutaneously. We
also need to restrict calls to mctp_hostlpc_kcs_send to be one at time.
This function is polling the KCS status register to ensure the IBF bit
is clear, and if it is will write the next KCS message to the IDR.
This entire function (the polling, and the setting of IDR) needs to be
done within a mutex block. mctp_hostlpc_kcs_send gets written 3 different
situations. 1.) Host notifies BMC a new message is ready for them to read,
2.) Host notifies BMC we have read their message, 3.) Host notifies
the BMC we wish to activate the channel. The mutex used in mctprp prior
to this commit was already protecting us for 1.) but we also need to protect
ourselves for situations 2.) and 3.) which this commit does.

Change-Id: Ia20c1d6bd904c1fb1a4bb2631def2baffda9cadf
CQ: SW535187
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123012
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>