Force Update of Minimum Secure Version to 2 for Early Ship Customers
This commit looks for and handles a special case in regards to
the system's Minimum Secure Version (MSV):
- If the MSV lockin policy is false --AND--
- The current Minimum Secure Version is < 2 -- AND --
- Secureboot is enabled...
Then set the minimum secure version to 2.

This is considered a special case because the MSV lockin policy is
being overruled and the value of 2 is not necessarily the exact
incoming MSV value.

This change is necessary for early ship customers who might get a
driver with SV==0, possibly loaded the initial GA driver of SV==1,
but we want to automatically update them to at least 2 once they get
an official production build with necessary security changes without
making them use the ASMI "Lockin Policy" option.

Change-Id: I8c83ddb59b74c64ff0ea6484162f33347294674b
CQ:SW535469
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123038
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Christopher J Engel <cjengel@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>