Create new error for SPI mux mismatches
There is an error path where the SPI mux can be flipped from PIB to FSI
during hostboot execution. Should hostboot attempt to access the mux
then a generic xscom error with a PIB resource occupied error will be
created and it calls out the proc. Instead, this commit will detect that
xscom error and callout the software instead of the proc since it's not
the hardware at fault.

Change-Id: I2a564d7abda32ed6e8426b973540303741f0a693
CQ:SW531888
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123106
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>