Support more than 255 Host PDRs
Previously, Hostboot was limited to notifying the BMC that up to 255
PDRs had been added to the host's PDR repository during the PDR
exchange, because of the encoding of the PLDM request event data. This
limit isn't large enough to accomodate fully equipped 8-processor
systems.

To get around this limitation, this commit alters the PDR exchange to
only notify the BMC about the first host PDR that was added to the
repository. The behavior of the BMC is to request that PDR and every
subsequent PDR in the repo, which allows an unlimited number of PDRs
to be sent to the BMC without explicitly notifying it of them.

Change-Id: I1f61577abd19cf6ea8e681bca0661711d37c9e8a
CQ: SW535331
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123146
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>