Standardize how Hostboot parses PLDM Bios attributes in host_set_ipl_parms
Prior to this commit, things were starting to get out of hand in the istep
host_set_ipl_parms. In this step Hostboot reads many attributes from the BMC
via PLDM commands. As more and more of these attributes have been added to
the 'list of BIOS attributes to parse' the code in the istep file has started
to turn chaotic. Some of these attributes had their own functions to call the
PLDM API and handle the value returned, others did not, and the functions that
were being written were going to the top of the file in host_set_ipl_params
and their interfaces were not consistent. To bring some order here this commit
introduces a new subdir, src/usr/isteps/bios_attr_accessors, which will contain
functions with matching APIs that can been looped through in host_set_ipl_parms.
In the future, when new BIOS attributes are added, a new function following this
API schema will be added bios_attr_parsers.C/H and added to the list of parsers
in host_set_ipl_parms.C.

Change-Id: Id2c03f32d7de2128db2fc60d66f3cc978ac07cf3
RTC: 288386
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123180
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>