WOF selection should look at present cores
As part of a change to handle ECO cores correctly, a problem was
introduced where we started looking for ALL cores (everything in
the blueprint) instead of PRESENT cores as we intended.

The fix is to modify the utility function to do what all the
callers assumed it was already doing.

Change-Id: I6f1f394c927c332666b26a5b296c2f7f447d23c6
CQ: SW535654
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123193
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>