Fix for SW534270
As a result of the fix for SW532974, transient TOD step check
errors during the TOD MPIPL configuration can lead to an
unintended checkstop.  The fix here is to move the disable
of the TOD step checkers to the beginning of the TOD MPIPL
configuration, to prevent the unintended checkstop.

Change-Id: I14000a0597dd15515bbd9750ee6b9d27c92b66a1
CQ:SW534270
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123239
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: HWSV CI <hwsv-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Benjamin Gass <bgass@us.ibm.com>
Reviewed-by: Nicholas Landi <nlandi@ibm.com>
Reviewed-by: Thi N Tran <thi@us.ibm.com>
Reviewed-by: DAVID A KING <kingboy@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123245
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>