Fix PLDM FRU table size calculation
The BMC includes a 4-byte CRC at the end of its response to PLDM FRU
Table fetch requests. Hostboot was not taking this into account when
allocating space for the response, so four bytes from the response
were overflowing the heap buffer and possibly causing some of the
crashes we have been seeing.

This commit fixes the issue by adding space for the CRC in the buffer
size calculation, and re-enables the size check.

Change-Id: I62c7bc9612cfc085999f193c28d83302ceaec25a
CQ: SW530637
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123464
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>