Set ATTR_DEFAULT_HB_CORE after applying gard records
This commit sets each proc's ATTR_DEFAULT_HB_CORE to
the ATTR_CHIP_UNIT of the first functional non-ECO
core on the proc. This value will later be written
to the customized SBE img. This is so the SBE can
choose an appropriate core to boot from on reconfig
loops if any cores are deconfigured during ipl.

Change-Id: I022985af40d3e9853fa2558af4102961685aaf1c
RTC: 281864
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123471
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>