Refactor update_pldm_subtree.sh to handle different subtrees
This commit renames update_pldm_subtree.sh to update_subtree.sh
and adds a required parameter to the script to specify the subtree
to update. Currently there are two subtree this tool supports,
libmctp and pldm. This commit also addresses a problem where
sometimes after merging the existing subtree with the latest upstream
repository there was lingering unmerged files if git thought that
both entities were 'adding' the file.

Change-Id: I0bd3f25f3288f4e9b86b5ee43e43a6f119b8811e
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123689
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>