add DONTCARE for container when searching for chassisoff effecter
This commit adds ENTITY_ID_DONTCARE for the entity_container_id when searching
for bmc_chassisoff_effecter_id. Before, this value was defaulted to 0, so
when the BMC changed their PDR entity ids, this value became out of sync.
There should only be one chassis entity, so we can use the ENTITY_ID_DONTCARE
constant there for the container.

Change-Id: Ia97bb0bc12cf5165351cdcab8b61a6389e9d9cc5
CQ:SW537056
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123774
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>