Fix bugs in hb_bios_attrs.C
This commit addresses two problems in hb_bios_attrs.C. The first
bug is that the BMC is starting to mark BIOS attributes it cannot
write as READ_ONLY. This changes the 'type' of the bios attribute.
Hostboot does not care if an attribute is type INTEGER or if it is
INTEGER_READ_ONLY, as the BMC is the only one that honors the read-
only status. The second bug is that Hostboot was writing integer
values in big-endian and it needs to write them in little-endian.

Change-Id: Ic54d0569a80c02ef5645e7dd4d4a22f9139a3f2a
RTC: 288386
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123936
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>