Fix bug in preverified lid manager
When loading lids without a secure header, the lid manager is
supposed to prepend a 4KB page as a 'fake' header to simplify
runtime logic that consumes the lids from reserved memory. The
logic that determines if this fake header is required or not was
inverted. This resulted in the lid load overrun the end of the
RINGOVD image in PNOR when hostboot loads it. Prior to using
PLDM for PNOR I/O the pnor section's VMM allocations were in
contiguous memory, however when we allocate VMM for the lids used
for PLDM PNOR the memory is not contiguous. This led to data
storage exceptions when we overran the buffer.

Change-Id: I00bb6663659d8455478249a63b5095fcb62aa7a4
RTC: 208802
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/123978
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>