Ignore READONLY bit for PLDM BIOS attributes is both directions
Prior to this commit if Hostboot requested a BIOS attribute of
type enum, and the BMC returned an attribute with type enum-readonly,
then hostboot would allow this descrepency. However if hostboot
expected enum-readonly and only enum was returned we would treat
it as an error. This commit makes the later case to be allowed also.
This statement is not only true for enum/enum-ro types but also the
only pldm bios attribute types : string/string-ro, integer/integer-ro,
and so on.

Change-Id: Iaaf75665b98c56d1b4f873542914dba570402ed6
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124158
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Michael Baiocchi <mbaiocch@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>