MDS: Code for MDS Media Controller presence detection
Added method fapiWrapif.H::get_mds_dev_addr that retrieves
the device address for the MDS for a given OCMB target.

Current method fapiWrapif.H::get_pmic_dev_addr queries the OCMB for
an SPD.  This is also needed by the fapiWrapif.H::get_mds_dev_addr
method therefore this code has been abstracted out into
a method of it's own and called by methods fapiWrapif.H::get_pmic_dev_addr
and fapiWrapif.H::get_mds_dev_addr.

Removed parameter i_buflen from method
i2cTargetPres.C::genericI2CTargetPresenceDetect.  It was not used
in the method but only to validate it's size.  The validation seems more
appropriate for the caller to do and should not be the function of the
callee.

And on that note, added an assertion of the input buffer length
to the methods that called genericI2CTargetPresenceDetect.  The original
validation of this parameter was to produce a firmware error. An incorrect
input size is not a firmware error but a programming error therefore I
believe an assertion is more appropriate.

Change-Id: Ia6ab584e882c6e43062d8fe6fb236378d0efa21d
RTC: 286322
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124172
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>