Support inhibiting BMC resets at runtime
This commit adds support for instructing the hypervisor on BMC systems
to avoid performing BMC resets at runtime, which will allow developers
to preserve the state of the system for debugging when an error
occurs. The PLDM BIOS attribute hb_inhibit_bmc_reset controls whether
Hostboot will set a field in HDAT to communicate the setting to the
hypervisor.

Change-Id: I1a7b65927a6d2640a4f8535f82a79586e267b38e
RTC: 293620
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124394
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>