MDS: Added code that will allow an explicit code update on MDS targets only
Added another field, MDS_FORCE_UPDATE, to the enumeration
OCMB_FW_UPDATE_BEHAVIOR that resides in the attribute_types.xml.

Added method isMdsDdimm that will be used in conjunction with
the field MDS_FORCE_UPDATE to check the SPD of the OCMB to determine
if the memory type is MDS and if so, do an update if MDS_FORCE_UPDATE
is set.

Change-Id: Ib2613af64125d5373fd2eeda710602c910625cf2
RTC: 284743
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124411
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>