Update Persistent HBD Partition
This commit introduces logic to merge the updated RW attributes from HBD
to the persistent HBD partition so that the persistent partition can be
updated with new RW attributes if there are any and old attributes with
size or persistency changes can be removed from the persistent partition.
The logic walks through all targets and RW attributes of the HBD
partition and copies their data over to the persistent partition in
expected format.

Change-Id: I4641a85184b34cdb6d1ccbcae1d8492568522d25
RTC: 205059
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124661
Reviewed-by: Christian R Geddes <crgeddes@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>