Fix std::vector constructor overload resolution
PROBLEM

std::vector has a constructor that takes two iterator parameters, one
to the beginning and one to the end of a sequence of values, and
constructs a vector from the elements in the range.

Previously this constructor was suitable for overload resolution for
any two arguments of the same type, regardless of whether they were
actually usable as iterators. This would cause compile failures in
cases like this, for example:

  std::vector<int> vec(1, 2);

This should construct a vector with one element being the value 2. Our
implementation would attempt to treat the values 1 and 2 as iterators
and fail the compilation.

SOLUTION

This commit adds qualifications to the signature of the iterator
constructor, so that it will only be available for overload resolution
when the type of the arguments supports an iterator interface.

Change-Id: I60aea2f9918529b4af24b4abf7265cfde8aff2cc
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/124568
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Nicholas E Bofferding <bofferdn@us.ibm.com>